Divided the code into modules, created modules
